+++ src/game/tictactoe.rs
@@ -61,7 +61,7 @@ impl game::Game for TicTacToe {
         self.history.push((player, *action));
     }
 
-    fn get_reward(&self) -> Option<f64> {
+    fn get_reward(&self) -> Option<f32> {
         
         match check_rows(&self)
             .or_else(|| check_cols(&self))
